+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#156](https://github.com/os2display/display-templates/pull/156)
+  - Clarified template creation.
 - [#155](https://github.com/os2display/display-templates/pull/155)
   - Added single-booking layout to calendar template.
   - Removed defaultProps.